feat: 每日自动派发兑换券

+++ common-exchange/src/main/kotlin/ink/pmc/common/exchange/ExchangeConfig.kt
@@ -19,7 +19,7 @@ object ExchangeConfig {
     }
 
     object Tickets {
-        var daily: Int = 0
+        var daily: Long = 0
     }
 
     object AvailableItems {
